- retour de la gestion des scores grâce à $id_jeu : ça ne fonctionnait plus depuis qq temps... - apparition des options du jeu et de la configuration interne dans jeux_voir.html - passage en fond d'une partie de jeux_edit.php - qq pas dans la compatibilité SPIP 2.0/SQL - qcm et mots-croisés : ajout des résultats détaillés - un truc que j'ai pas encore trop compris : le compactage...
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@26427 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _stable_/jeux/base/jeux_supprimer.php
@@ -1,11 +1,22 @@
 <?php
 function jeux_supprimer_tout_jeu($id_jeu){
-	spip_query('DELETE FROM spip_jeux_resultats WHERE id_jeu='.$id_jeu);
-	}
+	if(function_exists('sql_delete'))
+		sql_delete('spip_jeux_resultats', "id_jeu=$id_jeu");
+	else
+		spip_query("DELETE FROM spip_jeux_resultats WHERE id_jeu=$id_jeu");
+}
+
 function jeux_supprimer_tout_auteur($id_auteur){
-	spip_query('DELETE FROM spip_jeux_resultats WHERE id_auteur='.$id_auteur);
-	}
+	if(function_exists('sql_delete'))
+		sql_delete('spip_jeux_resultats', "id_auteur=$id_auteur");
+	else
+		spip_query("DELETE FROM spip_jeux_resultats WHERE id_auteur=$id_auteur");
+}
+
 function jeux_supprimer_tout_tout(){
-	spip_query('DELETE FROM spip_jeux_resultats');
-	}
+	if(function_exists('sql_delete'))
+		sql_delete('spip_jeux_resultats');
+	else
+		spip_query('DELETE FROM spip_jeux_resultats');
+}
 ?>
